pyproject.toml
[project]
name = "python-pro-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "just>=0.8.165",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.2",
]

# --- RUFF CONFIGURATION ---
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# E: pycodestyle, F: Pyflakes, I: isort, B: flake8-bugbear, UP: pyupgrade, N: pep8-naming
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["T201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- MYPY CONFIGURATION ---
[tool.mypy]
python_version = "3.11"
strict = true
show_error_codes = true
warn_unused_configs = true
# Essential for Conda: ignore libs without stubs so mypy doesn't crash on niche ML libs
ignore_missing_imports = true 

# Don't enforce strict typing in tests
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# --- PYTEST CONFIGURATION ---
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--strict-markers --cov=src --cov-report=term-missing"
log_cli = true

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
]
